Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char o;
-    scanf("%f %f %c",&a,&b,&o);
+    scanf("%d %d %c",&a,&b,&o);
     switch(o){
         case '+':
             printf("%d",a+b);
